refactor(engines): registry-driven /api/run + unified engine routes + kohya builtin pack (#301 #302)
/api/run now resolves model_train_type through the engine registry and
delegates to the owning pack's run.handle_run; the musubi/anima-fast
if-branches and the kohya privileged fall-through are removed from api.py
(1780 -> 819 lines). Shared submission helpers (sample prompts, sanitize,
SDXL prediction type, Anima defaults, tokenizer cache) move verbatim to
mikazuki/app/train_submit.py and are re-exported from api.py for
compatibility. Behavior of every pipeline stage is unchanged.

kohya becomes a builtin thin pack (mikazuki/engines/kohya/): manifest with
the real trainer_mapping capability matrix, TRAINER_MAPPING + the legacy
run block moved into the pack, status route reporting constant-ready.

Plugin routes unify to /api/engines/<id>/{status,preflight,dry-run,install,
repair,uninstall,install/log|progress/stream} via a generic dispatcher that
404s unknown engines and rejects install on builtin packs; the old
/api/plugins/* and /api/anima-fast/* compat routes are removed. Frontend
switches to the new routes and reads kohya status from the unified
endpoint (UI unchanged). Install-task log/progress stream URLs and the
legacy dist patch script are repointed accordingly.
