+++ frontend/src/api/training.ts
@@ -51,6 +51,6 @@ export const trainingApi = {
   validateImport: (pageTrainType: string, config: FormModel) => post<ImportValidation>("/api/config/validate-import", { page_train_type: pageTrainType, config }),
   normalizeExport: (pageTrainType: string, config: FormModel) => post<NormalizedExport>("/api/config/normalize-for-export", { page_train_type: pageTrainType, config }),
   run: (config: FormModel) => post<TrainingStart>("/api/run", config),
-  animaFastPreflight: (config: FormModel) => post<PreflightResult>("/api/anima-fast/preflight", config),
-  musubiPreflight: (config: FormModel) => post<PreflightResult>("/api/plugins/musubi/preflight", config),
+  animaFastPreflight: (config: FormModel) => post<PreflightResult>("/api/engines/anima-fast/preflight", config),
+  musubiPreflight: (config: FormModel) => post<PreflightResult>("/api/engines/musubi/preflight", config),
 }
