+++ frontend/src/pages/AnimaFastPage.vue
@@ -49,7 +49,7 @@ function closeStreams() {
 
 function streams(taskId: string, logUrl?: string, progressUrl?: string) {
   closeStreams()
-  logSource = new EventSource(logUrl || `/api/plugins/anima-lora/install/log/stream/${taskId}`)
+  logSource = new EventSource(logUrl || `/api/engines/anima-fast/install/log/stream/${taskId}`)
   logSource.onmessage = (event) => {
     try {
       const data = JSON.parse(event.data)
@@ -59,7 +59,7 @@ function streams(taskId: string, logUrl?: string, progressUrl?: string) {
       logs.value.push(event.data)
     }
   }
-  progressSource = new EventSource(progressUrl || `/api/plugins/anima-lora/install/progress/stream/${taskId}`)
+  progressSource = new EventSource(progressUrl || `/api/engines/anima-fast/install/progress/stream/${taskId}`)
   progressSource.onmessage = (event) => {
     try {
       const data = JSON.parse(event.data)
