+++ mikazuki/engines/kohya/__init__.py
@@ -1,0 +1,1 @@
+"""kohya (built-in sd-scripts) engine pack: thin wrapper over trainer_mapping."""
