+++ scripts/patch-anima-fast-entry.py
@@ -191,7 +191,7 @@ FAST_INSTALL_LOG_STYLE = (
     "line-height:1.45;white-space:pre-wrap;"
 )
 
-INSTALL_GUARD = r''';(()=>{if(window.__ANIMA_FAST_INSTALL_GUARD__)return;window.__ANIMA_FAST_INSTALL_GUARD__=true;const CONFIRM="Anima Fast 为进阶实验插件，需 NVIDIA GPU、约 16GB+ 显存，并会下载独立 Python 环境（数 GB）。\n\n确认已了解并继续安装？";let last={feature_enabled:true,state:"unknown"},es=null,tmr=null,scheduled=false;function q(s){return Array.from(document.querySelectorAll(s))}function isFastPage(){return/^\/lora\/anima-fast(\.html|\.md)?$/.test(location.pathname)}function markPage(){document.body.classList.toggle("anima-fast-page",isFastPage())}function setControls(d){if(!isFastPage())return;const kill=!d.feature_enabled,working=d.state==="installing"||d.state==="auditing",ready=d.state==="ready";q("[data-anima-fast-install]").forEach(b=>{b.disabled=kill||working;b.setAttribute("aria-disabled",b.disabled?"true":"false")});q(".right-container button").forEach(b=>{const t=(b.textContent||"").trim();if(t==="开始训练"||t==="✨加载训练预设✨"||t==="导入配置文件"||t==="保存参数"){b.disabled=kill||!ready;b.setAttribute("aria-disabled",b.disabled?"true":"false")}});document.body.classList.toggle("anima-fast-disabled",kill||!ready)}function label(d){if(!d.feature_enabled)return"功能已关闭";return d.state==="ready"?"插件已就绪":d.state==="installing"?"安装中":d.state==="auditing"?"审计中":d.state==="broken"?"需修复":d.state==="installed_unverified"?"待审计":"进阶插件 · 待开启"}function appendLog(x){const p=document.querySelector("[data-anima-fast-log]");if(!p)return;p.hidden=false;p.textContent+=(p.textContent?"\n":"")+x;p.scrollTop=p.scrollHeight}function apply(d){last=d||last;setControls(last);const n=document.querySelector("[data-anima-fast-status]");if(n)n.textContent=label(last);const a=last.facts&&last.facts.audit;if(a&&!a.ok&&a.errors)appendLog("[audit] "+a.errors.join("; "))}async function status(){try{const r=await fetch("/api/plugins/anima-lora/status"),j=await r.json();apply(Object.assign({feature_enabled:true},j.data||{state:"unknown"}))}catch(e){const n=document.querySelector("[data-anima-fast-status]");if(n)n.textContent="状态检查失败"}}function scheduleStatus(){if(scheduled)return;scheduled=true;setTimeout(()=>{scheduled=false;status()},120)}function openLog(url){if(!url||!window.EventSource)return;if(es)es.close();appendLog("[log] streaming "+url);es=new EventSource(url);es.onmessage=e=>{try{const d=JSON.parse(e.data);if(d.text)appendLog(d.text);if(d.done){appendLog("[log] done");es.close();es=null;if(tmr){clearInterval(tmr);tmr=null}status()}}catch(_){appendLog(e.data)}};es.onerror=()=>{appendLog("[log] stream disconnected");if(es){es.close();es=null}status()}}document.addEventListener("click",async e=>{const t=e.target&&e.target.closest&&e.target.closest("[data-anima-fast-guide-toggle]");if(t&&isFastPage()){const p=t.closest(".anima-fast-guide-collapsible"),b=p&&p.querySelector(".anima-fast-dataset-guide__body");if(b){const o=b.hidden;b.hidden=!o;t.setAttribute("aria-expanded",o?"true":"false");p.classList.toggle("is-open",o);try{localStorage.setItem("anima-fast-guide-open",o?"1":"0")}catch(_){}}return}const b=e.target&&e.target.closest&&e.target.closest("[data-anima-fast-install]");if(!b||!isFastPage())return;if(!last.feature_enabled)return;if(!window.confirm(CONFIRM))return;b.disabled=true;const s=document.querySelector("[data-anima-fast-status]"),p=document.querySelector("[data-anima-fast-log]");if(p){p.hidden=false;p.textContent=""}if(s)s.textContent="安装任务启动中";try{const r=await fetch("/api/plugins/anima-lora/install",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({dry_run:false})}),j=await r.json();if(j.status!=="success"){if(s)s.textContent=j.message||"安装失败";appendLog("[error] "+(j.message||"install failed"));return}const d=j.data||{};if(s)s.textContent="安装中";appendLog("[task] "+(d.task_id||"unknown"));openLog(d.log_stream||d.log_stream_url||(d.task_id?"/api/plugins/anima-lora/install/log/stream/"+d.task_id:""));if(tmr)clearInterval(tmr);tmr=setInterval(status,2000);status()}catch(t){if(s)s.textContent="安装失败";appendLog("[error] "+t)}finally{setTimeout(()=>setControls(last),250)}});function initGuideToggle(){if(!isFastPage())return;q("[data-anima-fast-guide-toggle]").forEach(t=>{const p=t.closest(".anima-fast-guide-collapsible"),b=p&&p.querySelector(".anima-fast-dataset-guide__body");if(!b)return;let o=false;try{o=localStorage.getItem("anima-fast-guide-open")==="1"}catch(_){}b.hidden=!o;t.setAttribute("aria-expanded",o?"true":"false");p.classList.toggle("is-open",o)})}new MutationObserver(scheduleStatus).observe(document.documentElement,{childList:true,subtree:true});document.addEventListener("DOMContentLoaded",()=>{markPage();initGuideToggle();status()});markPage();initGuideToggle();setTimeout(status,0)})();'''
+INSTALL_GUARD = r''';(()=>{if(window.__ANIMA_FAST_INSTALL_GUARD__)return;window.__ANIMA_FAST_INSTALL_GUARD__=true;const CONFIRM="Anima Fast 为进阶实验插件，需 NVIDIA GPU、约 16GB+ 显存，并会下载独立 Python 环境（数 GB）。\n\n确认已了解并继续安装？";let last={feature_enabled:true,state:"unknown"},es=null,tmr=null,scheduled=false;function q(s){return Array.from(document.querySelectorAll(s))}function isFastPage(){return/^\/lora\/anima-fast(\.html|\.md)?$/.test(location.pathname)}function markPage(){document.body.classList.toggle("anima-fast-page",isFastPage())}function setControls(d){if(!isFastPage())return;const kill=!d.feature_enabled,working=d.state==="installing"||d.state==="auditing",ready=d.state==="ready";q("[data-anima-fast-install]").forEach(b=>{b.disabled=kill||working;b.setAttribute("aria-disabled",b.disabled?"true":"false")});q(".right-container button").forEach(b=>{const t=(b.textContent||"").trim();if(t==="开始训练"||t==="✨加载训练预设✨"||t==="导入配置文件"||t==="保存参数"){b.disabled=kill||!ready;b.setAttribute("aria-disabled",b.disabled?"true":"false")}});document.body.classList.toggle("anima-fast-disabled",kill||!ready)}function label(d){if(!d.feature_enabled)return"功能已关闭";return d.state==="ready"?"插件已就绪":d.state==="installing"?"安装中":d.state==="auditing"?"审计中":d.state==="broken"?"需修复":d.state==="installed_unverified"?"待审计":"进阶插件 · 待开启"}function appendLog(x){const p=document.querySelector("[data-anima-fast-log]");if(!p)return;p.hidden=false;p.textContent+=(p.textContent?"\n":"")+x;p.scrollTop=p.scrollHeight}function apply(d){last=d||last;setControls(last);const n=document.querySelector("[data-anima-fast-status]");if(n)n.textContent=label(last);const a=last.facts&&last.facts.audit;if(a&&!a.ok&&a.errors)appendLog("[audit] "+a.errors.join("; "))}async function status(){try{const r=await fetch("/api/engines/anima-fast/status"),j=await r.json();apply(Object.assign({feature_enabled:true},j.data||{state:"unknown"}))}catch(e){const n=document.querySelector("[data-anima-fast-status]");if(n)n.textContent="状态检查失败"}}function scheduleStatus(){if(scheduled)return;scheduled=true;setTimeout(()=>{scheduled=false;status()},120)}function openLog(url){if(!url||!window.EventSource)return;if(es)es.close();appendLog("[log] streaming "+url);es=new EventSource(url);es.onmessage=e=>{try{const d=JSON.parse(e.data);if(d.text)appendLog(d.text);if(d.done){appendLog("[log] done");es.close();es=null;if(tmr){clearInterval(tmr);tmr=null}status()}}catch(_){appendLog(e.data)}};es.onerror=()=>{appendLog("[log] stream disconnected");if(es){es.close();es=null}status()}}document.addEventListener("click",async e=>{const t=e.target&&e.target.closest&&e.target.closest("[data-anima-fast-guide-toggle]");if(t&&isFastPage()){const p=t.closest(".anima-fast-guide-collapsible"),b=p&&p.querySelector(".anima-fast-dataset-guide__body");if(b){const o=b.hidden;b.hidden=!o;t.setAttribute("aria-expanded",o?"true":"false");p.classList.toggle("is-open",o);try{localStorage.setItem("anima-fast-guide-open",o?"1":"0")}catch(_){}}return}const b=e.target&&e.target.closest&&e.target.closest("[data-anima-fast-install]");if(!b||!isFastPage())return;if(!last.feature_enabled)return;if(!window.confirm(CONFIRM))return;b.disabled=true;const s=document.querySelector("[data-anima-fast-status]"),p=document.querySelector("[data-anima-fast-log]");if(p){p.hidden=false;p.textContent=""}if(s)s.textContent="安装任务启动中";try{const r=await fetch("/api/engines/anima-fast/install",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({dry_run:false})}),j=await r.json();if(j.status!=="success"){if(s)s.textContent=j.message||"安装失败";appendLog("[error] "+(j.message||"install failed"));return}const d=j.data||{};if(s)s.textContent="安装中";appendLog("[task] "+(d.task_id||"unknown"));openLog(d.log_stream||d.log_stream_url||(d.task_id?"/api/engines/anima-fast/install/log/stream/"+d.task_id:""));if(tmr)clearInterval(tmr);tmr=setInterval(status,2000);status()}catch(t){if(s)s.textContent="安装失败";appendLog("[error] "+t)}finally{setTimeout(()=>setControls(last),250)}});function initGuideToggle(){if(!isFastPage())return;q("[data-anima-fast-guide-toggle]").forEach(t=>{const p=t.closest(".anima-fast-guide-collapsible"),b=p&&p.querySelector(".anima-fast-dataset-guide__body");if(!b)return;let o=false;try{o=localStorage.getItem("anima-fast-guide-open")==="1"}catch(_){}b.hidden=!o;t.setAttribute("aria-expanded",o?"true":"false");p.classList.toggle("is-open",o)})}new MutationObserver(scheduleStatus).observe(document.documentElement,{childList:true,subtree:true});document.addEventListener("DOMContentLoaded",()=>{markPage();initGuideToggle();status()});markPage();initGuideToggle();setTimeout(status,0)})();'''
 
 
 def _guide_html_for_vue() -> str:
